Size should be a bigint, otherwise we have overflows in large trees

+++ testcases/verification/datastructures/RedBlackTree.scala
@@ -14,7 +14,7 @@ object RedBlackTree {
     case Node(_, l, v, r) => content(l) ++ Set(v) ++ content(r)
   }
 
-  def size(t : Tree) : Int = t match {
+  def size(t : Tree) : BigInt = t match {
     case Empty() => 0
     case Node(_, l, v, r) => size(l) + 1 + size(r)
   }
